Start reacting to the keyboard position
And reduce use of footer buttons
